Major Overhaul done to simplify command creation
Due to all of the statements happenning in the run_command() function,
major changes were done to make it easier to add commands and
automatically recognize if they have been typed and what arguments the
user provided. This is what the command.py class provides.

Permissions and any command related things were moved into their own
modules in the 'src/user' python package.

Lastly, linting and heavy commenting was done in all of the files to
make it easy to read and understand.