Added the ability to parse multiple words as one in a command
This can be done by setting the 3rd parameter in the CommandKeyword you are adding to KeywordCount.MULTIPLE_WORDS. Then using that keyword in your command. Note that it stops parsing at the '|' character or the end of what the user typed. The delimiter character '|' is used if other Keywords are to be used after.